Move transform code to seperate folder
