+++ Titanium-Alloy/src/Voxel/Voxel.h
@@ -15,6 +15,7 @@
 
 #include "src/Renderer/Renderer.h"
 #include "src/Renderer/Camera.h"
-#include "src/Renderer/Transform.h"
+
+#include "src/Transform/Transform.h"
 
 #include "src/Window/Window.h"
