+++ Titanium-Alloy/src/Voxel/src/Mesh/Mesh.h
@@ -9,7 +9,7 @@
 
 #include "MeshData.h"
 
-#include "../Renderer/Transform.h"
+#include "../Transform/Transform.h"
 #include "../Animation/Animation.h"
